backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.0
requests==2.31.0
psycopg2-binary==2.9.7
